Move aggregation selector into app bar
